sdxprairie: meta-qti-wlan: Bring up on LE.UM.3.3.2
Add sdxprairie target for bring-up on LE.UM.3.3.2.

base_conditional is obsolete on Yocto 3.0 and using it will
trigger not defined build error.

Fix is to use oe.utils.conditional, which is also backward
compatible with pre-Yocto 3.0 projects.

Also wlan-conf recipe is originally in meta-qti-bsp of Yocto
component. However in Yocto 3.0, system-conf is removed and
needs to be migrated into meta-qti-wlan. To keep backwards
compatibility, rename system-conf to wlan-conf. If
meta-qti-wlan is with Yocto 2.0 (e.g sdxpoorwills),
system-conf recipe will still be used from meta-qti-bsp.
If meta-qti-wlan is with Yocto 3.0 (e.g. sdxprairie),
wlan-conf will be used instead.

Change-Id: I75c34ba80f08c0f67e2635f1fe723f9f841a99dc
CRs-Fixed: 2346010