meta-qti-wlan: Fix build issue in BUILD_TAG
In BUILD_TAG feature, the 'git reflog' is run in wlan work
directory to extrace the latest ChangeId. In yocto system,
the work directory is under poky and this operation can't
get the desired information. Run 'git log' instead of 'git
reflog' in wlan source directory because there's nothing in
reflog.

Change-Id: I6aff1784b8111b1a5570338c84d6cc19d1d154c4
Signed-off-by: Guisen Yang <guiseny@codeaurora.org>