qcacld: sign module using mod-signing-keys
mod-signing-keys recipe was recently added to make sure that we create
a unique pair of keys that can be reused across builds, even when
TMPDIR is removed. We need to update the recipe to use the new
keys.

While doing the change remove the support for old signing keys, which
are not supported by mod-signing-keys anyways.

Change-Id: If063fa35d83e28d0ee17e3ca7f6f8a42c2509fb7
Signed-off-by: Nicolas Dechesne <ndeche@codeaurora.org>