meta-qti-wlan: Enable CONFIG_MDM_PLATFORM for STA IPA
In current design of wlan driver, the IPA pipe will be enabled
for STA mode when the macro CONFIG_MDM_PLATFORM enabled, otherwise
the IPA pipe will not be enabled if not have other connection on
SAP side(SAP+STA mode).

Currently, the macro CONFIG_MDM_PLATFORM is not enabled and then the
STA IPA does not work on STA only mode and STA+SAP mode if there is
not any other wlan client connected to SAP.

So enable this macro CONFIG_MDM_PLATFORM to make the STA IPA works fine.

Change-Id: I528eb25ee527eb7df3bfcc546891b8a68634979a
CRs-Fixed: 3508738