meta-qti-wlan: Adding variant to distinguish perf or debug WLAN module
In build_module.sh, the VARIANT always defaults to debug-defconfig,
regardless of whether the perf or debug variant is chosen for the
vendor build. As a result, a debug Module.symvers is generated for
WLAN module, and the debug CRC of the symbol module_layout is
captured in wlan.ko.

Therefore, if the device has a perf build image flashed and
CONFIG_MODVERSIONS is enabled, attempting to insmod WLAN module
will result in a CRC mismatch for the symbol module_layout, leading
to an insmod failure.

To fix the issue, we will pass build VARIENT information from
vendor build to build_modle.sh in recipe to distinguish perf or
debug WLAN module.

Change-Id: Ic6b4dbfbac0ebedf46df6efe9875cc2e1662d9d2
(cherry picked from commit 048974d92ac6164cb39186c20d610a9d93c942db)