meta-qti-wlan: Start wlan_daemon.service in simple mode
On boot-up, the wlan service was starting before the post_boot service.
Due to its configuration, it was blocking the execution of post_boot,
causing the system to remain stuck in waiting state. This lead to the
device running only on gold cores (core 6-7), while silver core (core
0-5) remained inactive untill post_boot was completed,to avoid this
changing type oneshot to simple.

Change-Id: I36841b518a1c0a5d47b1a4bf31563f525df4a1b7